Update LOVE_livesplit.asl
You may have missed my comment here: https://github.com/neesi/autosplitters/pull/1
I should've added that to the previous commit description instead.

It looks like the framecounter pointer path changes every time the data.win file is updated unless you happen to be lucky, and it has been updated 6 times in the last 10 game versions. So instead of using a pointer path, I used aob injection to grab the addresses. This way of doing it would've broke the splitter only once in the last 10 versions (after a major update). I'm sure there is a better / simpler way and I'm all for it.

What I did is more or less the same as this: https://youtu.be/kpSRUJfaT1o?t=726