clean structure, add signature scanning (#1)
* clean structure, add signature scanning

sig scanning should remove all possibility of the script breaking with updates

* Update LOVE_livesplit.asl

* Update LOVE_livesplit.asl

It looks like the framecounter pointer path changes every time the data.win file is updated unless you happen to be lucky, and it has been updated 6 times in the last 10 game versions. So instead of using a pointer path, I used aob injection to grab the addresses. This way of doing it would've broke the splitter only once in the last 10 versions (after a major update). I'm sure there is a better / simpler way and I'm all for it.

* Update LOVE_livesplit.asl

Works with at least 10 latest LOVE versions. Didn't test older ones. The signatures work with the dev's other games as well.

* Update LOVE_livesplit.asl

* Update LOVE_livesplit.asl

* Update LOVE_livesplit.asl

* Update LOVE_livesplit.asl

* Update LOVE_livesplit.asl

Co-authored-by: neesi